fix: use int8_t return type

+++ include/zenoh-pico/transport/multicast/lease.h
@@ -18,8 +18,8 @@
 #include "zenoh-pico/transport/transport.h"
 
 int8_t _zp_multicast_send_keep_alive(_z_transport_multicast_t *ztm);
-int _zp_multicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task);
-int _zp_multicast_stop_lease_task(_z_transport_t *zt);
+int8_t _zp_multicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task);
+int8_t _zp_multicast_stop_lease_task(_z_transport_t *zt);
 void *_zp_multicast_lease_task(void *ztm_arg);  // The argument is void* to avoid incompatible pointer types in tasks
 
 #endif /* ZENOH_PICO_MULTICAST_LEASE_H */
