+++ src/transport/unicast/lease.c
@@ -29,7 +29,7 @@ int8_t _zp_unicast_send_keep_alive(_z_transport_unicast_t *ztu) {
     return ret;
 }
 
-int _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
+int8_t _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
     // Init memory
     (void)memset(task, 0, sizeof(_z_task_t));
     // Attach task
@@ -43,7 +43,7 @@ int _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_ta
     return _Z_RES_OK;
 }
 
-int _zp_unicast_stop_lease_task(_z_transport_t *zt) {
+int8_t _zp_unicast_stop_lease_task(_z_transport_t *zt) {
     zt->_transport._unicast._lease_task_running = false;
     return _Z_RES_OK;
 }
@@ -113,14 +113,14 @@ int8_t _zp_unicast_send_keep_alive(_z_transport_unicast_t *ztu) {
     return _Z_ERR_TRANSPORT_NOT_AVAILABLE;
 }
 
-int _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
+int8_t _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
     _ZP_UNUSED(zt);
     _ZP_UNUSED(attr);
     _ZP_UNUSED(task);
     return _Z_ERR_TRANSPORT_NOT_AVAILABLE;
 }
 
-int _zp_unicast_stop_lease_task(_z_transport_t *zt) {
+int8_t _zp_unicast_stop_lease_task(_z_transport_t *zt) {
     _ZP_UNUSED(zt);
     return _Z_ERR_TRANSPORT_NOT_AVAILABLE;
 }
